extract wave create preview routes
